refactor: use hooks file
