+++ extension/src/notification/Widget.tsx
@@ -10,19 +10,13 @@ import {
   Checkbox,
   FormControlLabel,
 } from "@mui/material";
-import { atom, useAtomValue, useSetAtom } from "jotai";
-import { FC, memo, useState, useEffect } from "react";
+import { FC, memo } from "react";
 
-import { readyAtom } from "../prototypes/view/states/app";
 import { WidgetContext } from "../shared/context/WidgetContext";
-import {
-  isEnableAtom,
-  contentAtom,
-  startTimeAtom,
-  finishTimeAtom,
-} from "../shared/states/environmentVariables";
 import { ViewMarkdownViewer } from "../shared/ui-components/common";
 import { PLATEAUVIEW_NOTIFICATION_DOM_ID } from "../shared/ui-components/common/ViewClickAwayListener";
+
+import { useNotificationLogic } from "./hooks/useNotificationLogic";
 
 type NotificationProps = {
   isEnable?: boolean;
@@ -41,92 +35,11 @@ type NotificationWidgetProps<NotificationProps> = {
 };
 
 type Props = NotificationWidgetProps<NotificationProps>;
-const showAtom = atom<boolean>(false);
-
-// Function to generate a hash from content
-async function generateContentId(content: string): Promise<string> {
-  const encoder = new TextEncoder();
-  const data = encoder.encode(content);
-  const hashBuffer = await crypto.subtle.digest("SHA-256", data);
-  const hashArray = Array.from(new Uint8Array(hashBuffer));
-  const hashHex = hashArray.map((byte) => byte.toString(16).padStart(2, "0")).join("");
-  return hashHex;
-}
 
 export const Widget: FC<Props> = memo(function WidgetPresenter({ widget }) {
-  const ready = useAtomValue(readyAtom);
-  const [visible, setVisible] = useState(true);
-  const [doNotShowAgain, setDoNotShowAgain] = useState<boolean>(false);
+  const { ready, visible, show, content, handleClose, handleCheckboxChange, doNotShowAgain } = useNotificationLogic();
 
-  const isEnable = useAtomValue(isEnableAtom);
-  const content = useAtomValue(contentAtom);
-  const startTime = useAtomValue(startTimeAtom);
-  const finishTime = useAtomValue(finishTimeAtom);
-  const show = useAtomValue(showAtom);
-  const setShow = useSetAtom(showAtom);
-
-  const [contentId, setContentId] = useState<string | null>(null);
-
-  useEffect(() => {
-    const updateContentId = async () => {
-      if (content) {
-        const newContentId = await generateContentId(content);
-        setContentId(newContentId);
-
-        const storedDoNotShowAgain = localStorage.getItem("doNotShowAgain");
-        const storedContentID = localStorage.getItem("contentId");
-
-        if (storedDoNotShowAgain === "true" && storedContentID === newContentId) {
-          setVisible(false);
-          return;
-        }
-      }
-    };
-    updateContentId();
-  }, [content]);
-
-  useEffect(() => {
-    // Get the current time and convert it to UTC
-    const nowUTC = new Date().getTime();
-
-    // Convert `startTime` and `finishTime` to UTC
-    const startTimeUTC = startTime ? new Date(startTime).getTime() : null;
-    const finishTimeUTC = finishTime ? new Date(finishTime).getTime() : null;
-
-    // Check display conditions
-    if (isEnable) {
-      if (!startTimeUTC && !finishTimeUTC) {
-        // If neither is set, always show
-        setShow(true);
-      } else if (startTimeUTC && !finishTimeUTC) {
-        // If only `startTime` is set, show if current time is after `startTime`
-        setShow(nowUTC >= startTimeUTC);
-      } else if (!startTimeUTC && finishTimeUTC) {
-        // If only `finishTime` is set, show if current time is before `finishTime`
-        setShow(nowUTC <= finishTimeUTC);
-      } else if (startTimeUTC && finishTimeUTC) {
-        // If both are set, show if current time is between `startTime` and `finishTime`
-        setShow(nowUTC >= startTimeUTC && nowUTC <= finishTimeUTC);
-      }
-    } else {
-      setShow(false); // Always hide if `isEnable` is false
-    }
-  }, [isEnable, startTime, finishTime, contentId, setShow]);
-
-  const handleClose = () => {
-    setVisible(false);
-    if (doNotShowAgain) {
-      // Save the "do not show again" flag to localStorage
-      localStorage.setItem("doNotShowAgain", "true");
-      localStorage.setItem("contentId", contentId || "");
-    }
-  };
-
-  const handleCheckboxChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setDoNotShowAgain(event.target.checked);
-  };
-
-  if (!visible) return null; // 非表示の場合、何もレンダリングしない
+  if (!visible) return null;
 
   return (
     <div id={PLATEAUVIEW_NOTIFICATION_DOM_ID}>
@@ -168,7 +81,7 @@ export const Widget: FC<Props> = memo(function WidgetPresenter({ widget }) {
               }}>
               <FormControlLabel
                 control={
-                  <Checkbox 
+                  <Checkbox
                     name="doNotShowAgain"
                     checked={doNotShowAgain}
                     onChange={handleCheckboxChange}
